+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Yii Framework 2 redis extension Change Log
 - Enh #281: PHP 7.3 support was removed to align with Yii2 minimum PHP version (@s1lver)
 - Enh #287: Applying Yii2 coding standards (@s1lver)
 - Fix #281: PHP 7.4 compatibility is now fixed (@s1lver)
+- Fix: Prevent infinite loop in `parseResponse()` when `fread()` returns empty string on broken socket (@dkostik)
 
 2.1.0 December 25, 2025
 -----------------------